+++ lib/dimp.dart
@@ -10,12 +10,14 @@ export 'src/protocol/contents.dart';
 export 'src/protocol/files.dart';
 export 'src/protocol/commands.dart';
 export 'src/protocol/groups.dart';
+export 'src/protocol/receipt.dart';
 
 export 'src/dkd/base.dart';
 export 'src/dkd/contents.dart';
 export 'src/dkd/files.dart';
 export 'src/dkd/commands.dart';
 export 'src/dkd/groups.dart';
+export 'src/dkd/receipt.dart';
 export 'src/dkd/factory.dart';
 
 export 'src/msg/base.dart';